Update game.html
adjusted the "coinDropInterval"and sequence patterns to make the gameplay more dynamic and challenging. 

i added some variety in the intervals and sequences to increase the difficulty as the player progresses. 

Varying Intervals:  (randomized coin drop interval within a certain range to make the timing less predictable)  

Progressive Difficulty: (As the player progresses through the track, the coin sequences will become more frequent and complex)